requirements.txt
fastapi
uvicorn
pandas
scikit-learn
python-dotenv
httpx
pydantic
numpy
openpyxl
